requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
PyPDF2==3.0.1
faiss-cpu==1.7.4
groq==0.3.1
sentence-transformers==2.2.2
huggingface_hub==0.16.4
langchain==0.1.14
langchain-community==0.0.30
transformers==4.29.2
